move the AbstractDbafsMetadataEvent class to the correct location
