+++ core-bundle/src/Filesystem/Dbafs/AbstractDbafsMetadataEvent.php
@@ -10,7 +10,7 @@ declare(strict_types=1);
  * @license LGPL-3.0-or-later
  */
 
-namespace Contao\CoreBundle\Event;
+namespace Contao\CoreBundle\Filesystem\Dbafs;
 
 use Symfony\Component\Uid\Uuid;
